--- src/model/BoundryBox.ts
@@ -1,4 +1,0 @@
-export default class BoundaryBox {
-  constructor (readonly top: number, readonly left: number, readonly width: number, readonly height: number) {
-  }
-}
