+++ src/model/PageBlock.ts
@@ -1,7 +1,7 @@
 import { CommonBlock, Blocks } from '../types'
 import ParentBlock from './ParentBlock'
 import LineBlock from './LineBlock'
-import Geometry from './Geometry'
+import { Geometry } from './Geometry'
 
 export default class PageBlock extends ParentBlock<LineBlock> implements CommonBlock {
   readonly blockType: 'PAGE' = 'PAGE'
