--- src/model/Point.ts
@@ -1,4 +1,0 @@
-export default class Point {
-  constructor (readonly x: number, readonly y: number) {
-  }
-}
