+++ src/types.ts
@@ -1,5 +1,4 @@
-import Point from './model/Point'
-import Geometry from './model/Geometry'
+import { Geometry } from './model/Geometry'
 import PageBlock from './model/PageBlock'
 import LineBlock from './model/LineBlock'
 import WordBlock from './model/WordBlock'
@@ -17,8 +16,6 @@ export interface TextProvider {
 export type Block = PageBlock | LineBlock | WordBlock
 
 export type Blocks<T extends Block> = T[]
-
-export type Polygon = Point[]
 
 export interface DocumentMetadata {
   readonly pages: number
